code review and fixes

+++ src/WebApi/Controllers/Admin/AdminGroupMemberController.cs
@@ -2,6 +2,7 @@ using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using WebApi.DTOs.Responses;
 using WebApi.DTOs.Responses.Admin.GroupMember;
+using WebApi.DTOs.Responses.Admin.GroupMemberDestinationVote;
 using WebApi.DTOs.Responses.GroupMemberDestinationVote;
 using WebApi.Persistence.Interfaces;
 using WebApi.Repositories.Implementations;
@@ -89,6 +90,11 @@ public class AdminGroupMemberController(
             return NotFound(new ErrorResponse("The user is not a member of this group."));
         }
 
+        if (userToRemove.IsOwner)
+        {
+            return BadRequest(new ErrorResponse("It is not possible to remove the owner of group."));
+        }
+
         groupMemberRepository.Remove(userToRemove);
         await unitOfWork.SaveAsync();
 
@@ -118,7 +124,7 @@ public class AdminGroupMemberController(
 
         return Ok(new AdminGetGroupMemberAllDestinationVotesByIdResponse
         {
-            Data = votes.Select(x => new AdminGetAllGroupDestinationVotesByIdResponseData
+            Data = votes.Select(x => new AdminGetGroupMemberAllDestinationVotesByIdResponseData
             {
                 Id = x.Id,
                 CreatedAt = x.CreatedAt
